+++ next-env.d.ts
@@ -3,3 +3,5 @@
 
 // NOTE: This file should not be edited
 // see https://nextjs.org/docs/basic-features/typescript for more information.
+
+declare module 'micro-method-router';
