+++ pages/api/auth/token.ts
@@ -1,9 +1,9 @@
-import type { NextApiRequest, NextApiResponse } from "next";
-import * as Yup from "yup";
+import type { NextApiRequest, NextApiResponse } from 'next';
+import * as Yup from 'yup';
 
-import methods from "micro-method-router";
-import { validationMiddleware } from "../../../middlewares";
-import { createToken } from "../../../controllers/order-controller";
+import methods from 'micro-method-router';
+import { validationMiddleware } from '../../../middlewares';
+import { createToken } from 'controllers/auth-controller';
 
 const schema = Yup.object()
   .shape({
@@ -17,7 +17,7 @@ export async function post(req: NextApiRequest, res: NextApiResponse) {
   const { email, code } = req.body;
   try {
     const { token } = await createToken(email, code);
-    if (!token) throw new Error("Token was not create");
+    if (!token) throw new Error('Token was not create');
 
     res.status(201).send({ error: null, token });
   } catch (error) {
